requirements.txt
asttokens==3.0.0
debugpy==1.8.13
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
ipython==8.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pydevd-odoo==1.2.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
